chore: rename RenderInputBox to RenderTextControl
