+++ kraken/lib/rendering.dart
@@ -17,6 +17,7 @@ export 'src/rendering/intrinsic.dart';
 export 'src/rendering/sliver_list.dart';
 export 'src/rendering/preferred_size.dart';
 export 'src/rendering/text.dart';
+export 'src/rendering/text_control.dart';
 export 'src/rendering/content_visibility.dart';
 export 'src/rendering/pointer.dart';
 export 'src/rendering/opacity.dart';
